Create a feature_registration feature flag
This feature flag will be used to open or to close registrations on the
server. By default, the flag is enabled to keep the previous behaviour.

The value is also created in the seeds. Because we make sure to remove
all previous flags (`FlipperFeature.destroy`), we need to create the
corresponding model too.