fix: #305 Fix Docker build
`openssl-dev` seems to not be needed anymore and was conflicting with
`libressl-dev` (dependency of `postgresql-dev`).

npm@5 was failing with `Error: write after end` error. It has been fixed
in npm@6 so we just upgrade. I cross my fingers. I'm so tired of NPM.
Seriously? Each time I want to work on this project there is something
wrong with NPM.