Use ProjectCardDeck component on projects page
I wanted to avoid the use of this component on this page because I think
it could be confusing with the dashboard page; but it appears that it's
confusing to NOT have it, so let's try. The reason of this confusion is
that when we access a project via the dashboard, the active link becomes
"Projects" instead of "Dashboard". So when we click on "Projects", we
expect to have the same view for our projects.